requirements.txt
langchain-huggingface
langchain-chroma
langchain-ollama
sentence-transformers
